refactor(acp): isolate timeout cleanup lifecycle
